requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
Django==2.2.1
django-groupadmin-users==0.3.3
-e git+https://github.com/Microdisseny/django-theme-microdisseny.git@f5115fbb38abc286fc5ff41e045f77d1c020609f#egg=django_theme_microdisseny
python-slugify==3.0.2
pytz==2019.1
requests==2.22.0
